requirements.txt
canoser==0.8.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
idna==2.10
mnemonic==0.19
pycparser==2.20
PyNaCl==1.4.0
PyQt5==5.15.1
PyQt5-sip==12.8.1
requests==2.24.0
six==1.15.0
urllib3==1.25.10
